Modified category controllers and view

+++ app/controllers/category_controller.rb
@@ -9,15 +9,27 @@ class CategoryController < ApplicationController
         erb :'category/index'
     end
 
-    post '/categories' do
+    post '/categories/add' do
+      if !exists_already?(params[:category_name])
+        binding.pry
         name = params[:category_name]
         cat = Category.create(:category_name => name)
         Category.all << cat
         cat.save
-
+      else
+        "Error, category already exists"
+      end
         redirect to '/categories'
     end
 
-    helpers do 
+    helpers do
+
+      def exists_already?(name)
+        name = Category.find_by(:category_name => name)
+        if name != nil
+          return true
+        end
+        return false
+      end
     end
 end
